+++ test/suite/Assertion/Renderer/AssertionRendererTest.php
@@ -18,6 +18,8 @@ use Eloquent\Phony\Event\NullEvent;
 use Eloquent\Phony\Exporter\InlineExporter;
 use Eloquent\Phony\Invocation\InvocableInspector;
 use Eloquent\Phony\Matcher\EqualToMatcher;
+use Eloquent\Phony\Mock\Builder\MockBuilder;
+use Eloquent\Phony\Mock\Proxy\Factory\ProxyFactory;
 use Eloquent\Phony\Spy\Spy;
 use Eloquent\Phony\Stub\Stub;
 use Eloquent\Phony\Test\TestCallFactory;
@@ -40,13 +42,25 @@ class AssertionRendererTest extends PHPUnit_Framework_TestCase
         $this->invocableInspector = new InvocableInspector();
         $this->exporter = new InlineExporter(false);
         $this->subject = new AssertionRenderer($this->invocableInspector, $this->exporter);
+        $this->proxyFactory = ProxyFactory::instance();
 
-        $this->thisObject = new TestClassA();
+        $this->thisObjectA = new TestClassA();
+
+        $mockBuilder = new MockBuilder('Eloquent\Phony\Test\TestClassA');
+        $this->thisObjectB = $mockBuilder->get();
+        $this->thisObjectBProxy = $this->proxyFactory->createStubbing($this->thisObjectB);
+        $this->thisObjectBProxy->setLabel('label');
+        $this->thisObjectB->testClassAMethodA();
+
+        $mockBuilder = new MockBuilder('IteratorAggregate', null, 'PhonyMockAssertionRendererTestIteratorAggregate');
+        $this->thisObjectC = $mockBuilder->get();
+        $this->thisObjectCProxy = $this->proxyFactory->createStubbing($this->thisObjectC);
+        $this->thisObjectCProxy->setLabel('label');
 
         $this->callFactory = new TestCallFactory();
         $this->callEventFactory = $this->callFactory->eventFactory();
         $this->callA = $this->callFactory->create(
-            $this->callEventFactory->createCalled(array($this->thisObject, 'testClassAMethodA'), array('a', 'b')),
+            $this->callEventFactory->createCalled(array($this->thisObjectA, 'testClassAMethodA'), array('a', 'b')),
             $this->callEventFactory->createReturned('x')
         );
         $this->callB = $this->callFactory->create(
@@ -56,6 +70,13 @@ class AssertionRendererTest extends PHPUnit_Framework_TestCase
         $this->callC = $this->callFactory->create(
             $this->callEventFactory->createCalled('implode')
         );
+        $this->callD = $this->callFactory->create(
+            $this->callEventFactory->createCalled(array($this->thisObjectB, 'testClassAMethodA'))
+        );
+        $this->callE = $this->callFactory->create(
+            $this->callEventFactory->createCalled(array($this->thisObjectC, 'getIterator'))
+        );
+        $this->callF = $this->thisObjectBProxy->testClassAMethodA->callAt();
     }
 
     public function testConstructor()
@@ -83,6 +104,26 @@ class AssertionRendererTest extends PHPUnit_Framework_TestCase
         );
     }
 
+    public function testRenderMock()
+    {
+        $this->assertSame('TestClassA[label]', $this->subject->renderMock($this->thisObjectBProxy));
+    }
+
+    public function testRenderMockStatic()
+    {
+        $proxy = $this->proxyFactory->createStubbingStatic($this->thisObjectB);
+
+        $this->assertSame('TestClassA[static]', $this->subject->renderMock($proxy));
+    }
+
+    public function testRenderMockWithoutParentClass()
+    {
+        $this->assertSame(
+            'PhonyMockAssertionRendererTestIteratorAggregate[label]',
+            $this->subject->renderMock($this->thisObjectCProxy)
+        );
+    }
+
     public function testRenderMatchers()
     {
         $matcherA = new EqualToMatcher('a');
@@ -93,63 +134,21 @@ class AssertionRendererTest extends PHPUnit_Framework_TestCase
         $this->assertSame('"a", 111', $this->subject->renderMatchers(array($matcherA, $matcherB)));
     }
 
-    public function renderCardinalityData()
-    {
-        //                                                minimum maximum isAlways verb      expected
-        return array(
-            'Null minimum, null maximum'         => array(null,   null,   false,   'return', 'return, any number of times'),
-            'Zero minimum, null maximum'         => array(0,      null,   false,   'return', 'return, any number of times'),
-            'One minimum, null maximum'          => array(1,      null,   false,   'return', 'return'),
-            'Two minimum, null maximum'          => array(2,      null,   false,   'return', 'return, 2 times'),
-
-            'Null minimum, zero maximum'         => array(null,   0,      false,   'return', 'no return'),
-            'Zero minimum, zero maximum'         => array(0,      0,      false,   'return', 'no return'),
-
-            'Null minimum, one maximum'          => array(null,   1,      false,   'return', 'return, up to 1 time'),
-            'Zero minimum, one maximum'          => array(0,      1,      false,   'return', 'return, up to 1 time'),
-            'One minimum, one maximum'           => array(1,      1,      false,   'return', 'return, exactly 1 time'),
-
-            'Null minimum, two maximum'          => array(null,   2,      false,   'return', 'return, up to 2 times'),
-            'Zero minimum, two maximum'          => array(0,      2,      false,   'return', 'return, up to 2 times'),
-            'One minimum, two maximum'           => array(1,      2,      false,   'return', 'return, between 1 and 2 times'),
-            'Two minimum, two maximum'           => array(2,      2,      false,   'return', 'return, exactly 2 times'),
-
-            'Null minimum, null maximum, always' => array(null,   null,   true,    'return', 'every return, any number of times'),
-            'Zero minimum, null maximum, always' => array(0,      null,   true,    'return', 'every return, any number of times'),
-            'One minimum, null maximum, always'  => array(1,      null,   true,    'return', 'every return'),
-            'Two minimum, null maximum, always'  => array(2,      null,   true,    'return', 'every return, 2 times'),
-
-            'Null minimum, one maximum, always'  => array(null,   1,      true,    'return', 'every return, up to 1 time'),
-            'Zero minimum, one maximum, always'  => array(0,      1,      true,    'return', 'every return, up to 1 time'),
-            'One minimum, one maximum, always'   => array(1,      1,      true,    'return', 'every return, exactly 1 time'),
-
-            'Null minimum, two maximum, always'  => array(null,   2,      true,    'return', 'every return, up to 2 times'),
-            'Zero minimum, two maximum, always'  => array(0,      2,      true,    'return', 'every return, up to 2 times'),
-            'One minimum, two maximum, always'   => array(1,      2,      true,    'return', 'every return, between 1 and 2 times'),
-            'Two minimum, two maximum, always'   => array(2,      2,      true,    'return', 'every return, exactly 2 times'),
-        );
-    }
-
-    /**
-     * @dataProvider renderCardinalityData
-     */
-    public function testRenderCardinality($minimum, $maximum, $isAlways, $verb, $expected)
-    {
-        $this->assertSame(
-            $expected,
-            $this->subject->renderCardinality(new Cardinality($minimum, $maximum, $isAlways), $verb)
-        );
-    }
-
     public function testRenderCalls()
     {
         $expected = <<<'EOD'
-    - Eloquent\Phony\Test\TestClassA->testClassAMethodA("a", "b")
+    - TestClassA->testClassAMethodA("a", "b")
     - implode()
+    - TestClassA->testClassAMethodA()
+    - IteratorAggregate->getIterator()
+    - TestClassA[label]->testClassAMethodA()
 EOD;
 
         $this->assertSame('', $this->subject->renderCalls(array()));
-        $this->assertSame($expected, $this->subject->renderCalls(array($this->callA, $this->callB)));
+        $this->assertSame(
+            $expected,
+            $this->subject->renderCalls(array($this->callA, $this->callB, $this->callD, $this->callE, $this->callF))
+        );
     }
 
     public function testRenderThisValues()
@@ -221,7 +220,7 @@ EOD;
         return array(
             'Method' => array(
                 $callFactory->create($callEventFactory->createCalled(array($this, 'setUp'))),
-                'Eloquent\Phony\Assertion\Renderer\AssertionRendererTest->setUp()',
+                'AssertionRendererTest->setUp()',
             ),
             'Static method' => array(
                 $callFactory->create($callEventFactory->createCalled('ReflectionMethod::export')),
@@ -326,9 +325,9 @@ EOD;
             )
         );
         $expected = <<<'EOD'
-    - called Eloquent\Phony\Test\TestClassA->testClassAMethodA("a", "b")
-    - called Eloquent\Phony\Test\TestClassA->testClassAMethodA("a", "b")
-    - returned "x" from Eloquent\Phony\Test\TestClassA->testClassAMethodA("a", "b")
+    - called TestClassA->testClassAMethodA("a", "b")
+    - called TestClassA->testClassAMethodA("a", "b")
+    - returned "x" from TestClassA->testClassAMethodA("a", "b")
     - threw RuntimeException("You done goofed.") in implode()
     - produced "x": "y" from unknown call
     - received "z" in unknown call
@@ -338,6 +337,54 @@ EOD;
 EOD;
 
         $this->assertSame($expected, $this->subject->renderEvents($events));
+    }
+
+    public function renderCardinalityData()
+    {
+        //                                                minimum maximum isAlways verb      expected
+        return array(
+            'Null minimum, null maximum'         => array(null,   null,   false,   'return', 'return, any number of times'),
+            'Zero minimum, null maximum'         => array(0,      null,   false,   'return', 'return, any number of times'),
+            'One minimum, null maximum'          => array(1,      null,   false,   'return', 'return'),
+            'Two minimum, null maximum'          => array(2,      null,   false,   'return', 'return, 2 times'),
+
+            'Null minimum, zero maximum'         => array(null,   0,      false,   'return', 'no return'),
+            'Zero minimum, zero maximum'         => array(0,      0,      false,   'return', 'no return'),
+
+            'Null minimum, one maximum'          => array(null,   1,      false,   'return', 'return, up to 1 time'),
+            'Zero minimum, one maximum'          => array(0,      1,      false,   'return', 'return, up to 1 time'),
+            'One minimum, one maximum'           => array(1,      1,      false,   'return', 'return, exactly 1 time'),
+
+            'Null minimum, two maximum'          => array(null,   2,      false,   'return', 'return, up to 2 times'),
+            'Zero minimum, two maximum'          => array(0,      2,      false,   'return', 'return, up to 2 times'),
+            'One minimum, two maximum'           => array(1,      2,      false,   'return', 'return, between 1 and 2 times'),
+            'Two minimum, two maximum'           => array(2,      2,      false,   'return', 'return, exactly 2 times'),
+
+            'Null minimum, null maximum, always' => array(null,   null,   true,    'return', 'every return, any number of times'),
+            'Zero minimum, null maximum, always' => array(0,      null,   true,    'return', 'every return, any number of times'),
+            'One minimum, null maximum, always'  => array(1,      null,   true,    'return', 'every return'),
+            'Two minimum, null maximum, always'  => array(2,      null,   true,    'return', 'every return, 2 times'),
+
+            'Null minimum, one maximum, always'  => array(null,   1,      true,    'return', 'every return, up to 1 time'),
+            'Zero minimum, one maximum, always'  => array(0,      1,      true,    'return', 'every return, up to 1 time'),
+            'One minimum, one maximum, always'   => array(1,      1,      true,    'return', 'every return, exactly 1 time'),
+
+            'Null minimum, two maximum, always'  => array(null,   2,      true,    'return', 'every return, up to 2 times'),
+            'Zero minimum, two maximum, always'  => array(0,      2,      true,    'return', 'every return, up to 2 times'),
+            'One minimum, two maximum, always'   => array(1,      2,      true,    'return', 'every return, between 1 and 2 times'),
+            'Two minimum, two maximum, always'   => array(2,      2,      true,    'return', 'every return, exactly 2 times'),
+        );
+    }
+
+    /**
+     * @dataProvider renderCardinalityData
+     */
+    public function testRenderCardinality($minimum, $maximum, $isAlways, $verb, $expected)
+    {
+        $this->assertSame(
+            $expected,
+            $this->subject->renderCardinality(new Cardinality($minimum, $maximum, $isAlways), $verb)
+        );
     }
 
     public function testInstance()