+++ test/suite/Mock/Proxy/Stubbing/StaticStubbingProxyTest.php
@@ -192,10 +192,10 @@ class StaticStubbingProxyTest extends PHPUnit_Framework_TestCase
         $className::testClassAStaticMethodB('c', 'd');
         $className::testClassAStaticMethodA('e', 'f');
         $expected = <<<'EOD'
-Expected no interaction with PhonyMockStaticStubbingNoInteraction[static]. Calls:
-    - PhonyMockStaticStubbingNoInteraction::testClassAStaticMethodA("a", "b")
-    - PhonyMockStaticStubbingNoInteraction::testClassAStaticMethodB("c", "d")
-    - PhonyMockStaticStubbingNoInteraction::testClassAStaticMethodA("e", "f")
+Expected no interaction with TestClassA[static]. Calls:
+    - TestClassA::testClassAStaticMethodA("a", "b")
+    - TestClassA::testClassAStaticMethodB("c", "d")
+    - TestClassA::testClassAStaticMethodA("e", "f")
 EOD;
 
         $this->setExpectedException('Eloquent\Phony\Assertion\Exception\AssertionException', $expected);